refactors message validator
