+++ cmd/ghost/main.go
@@ -6,7 +6,7 @@ import (
 	"github.com/onflow/flow-go/cmd"
 	"github.com/onflow/flow-go/engine/ghost/engine"
 	"github.com/onflow/flow-go/module"
-	"github.com/onflow/flow-go/network/validators"
+	"github.com/onflow/flow-go/network/validator"
 )
 
 func main() {
@@ -19,9 +19,9 @@ func main() {
 			flags.StringVarP(&rpcConf.ListenAddr, "rpc-addr", "r", "localhost:9000", "the address the GRPC server listens on")
 		}).
 		Module("message validators", func(node *cmd.FlowNodeBuilder) error {
-			node.MsgValidators = []validators.MessageValidator{
+			node.MsgValidators = []validator.MessageValidator{
 				// filter out messages sent by this node itself
-				validators.NewSenderValidator(node.Me.NodeID()),
+				validator.NewSenderValidator(node.Me.NodeID()),
 				// but retain all the 1-k messages even if they are not intended for this node
 			}
 			return nil
