+++ network/validator.go
@@ -1,4 +1,4 @@
-package validators
+package network
 
 import "github.com/onflow/flow-go/network/message"
